# backend/api-spec.json
{"openapi":"3.1.0","info":{"title":"Revele Cosméticos — API","description":"API do SaaS de assinaturas e vendas da Revele Cosméticos.","contact":{"name":"Luiz","email":"seu-email@exemplo.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/usuarios/{id}":{"put":{"tags":["user-controller"],"operationId":"putUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/planos/{id}":{"put":{"tags":["plano-controller"],"operationId":"putPlan","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanoResponseDTO"}}}}}},"delete":{"tags":["plano-controller"],"operationId":"deletePlan","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/usuarios/login":{"post":{"tags":["user-controller"],"operationId":"userLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}}}},"/usuarios/cadastrar":{"post":{"tags":["user-controller"],"operationId":"userRegister","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCadastroDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/produtos":{"get":{"tags":["produto-controller"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProdutoResponseDTO"}}}}}}},"post":{"tags":["produto-controller"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdutoCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoResponseDTO"}}}}}}},"/planos":{"post":{"tags":["plano-controller"],"operationId":"createPlan","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanoResponseDTO"}}}}}}},"/pedidos":{"post":{"tags":["pedido-controller"],"operationId":"criarPedido","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PedidoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponseDTO"}}}}}}},"/assinaturas/":{"post":{"tags":["assinatura-controller"],"operationId":"fazerAssinatura","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssinaturaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssinaturaResponseDTO"}}}}}}},"/produtos/{id}":{"get":{"tags":["produto-controller"],"operationId":"buscar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoResponseDTO"}}}}}},"delete":{"tags":["produto-controller"],"operationId":"deletar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["produto-controller"],"operationId":"modificar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdutoUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdutoResponseDTO"}}}}}}},"/assinaturas/{id}/renovar":{"patch":{"tags":["assinatura-controller"],"operationId":"renovarAssinatura","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssinaturaResponseDTO"}}}}}}},"/assinaturas/{id}/pausar":{"patch":{"tags":["assinatura-controller"],"operationId":"alterarStatusAssinatura","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssinaturaResponseDTO"}}}}}}},"/assinaturas/{id}/cancelar":{"patch":{"tags":["assinatura-controller"],"operationId":"cancelarAssinatura","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssinaturaResponseDTO"}}}}}}},"/usuarios/perfil":{"get":{"tags":["user-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/usuarios/admin/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}}},"/planos/":{"get":{"tags":["plano-controller"],"operationId":"listarPlanos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlanoResponseDTO"}}}}}}}},"/pedidos/{id}":{"get":{"tags":["pedido-controller"],"operationId":"buscarPedido","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponseDTO"}}}}}}},"/pedidos/meus":{"get":{"tags":["pedido-controller"],"operationId":"listarMeusPedidos","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagePedidoResponseDTO"}}}}}}},"/assinaturas/admin":{"get":{"tags":["assinatura-controller"],"operationId":"verAssinaturas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AssinaturaResponseDTO"}}}}}}}}},"components":{"schemas":{"UserUpdateDTO":{"type":"object","properties":{"email":{"type":"string","format":"email"},"name":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"name":{"type":"string"},"id":{"type":"string","format":"uuid"}}},"PlanoDTO":{"type":"object","properties":{"nome":{"type":"string"},"descricao":{"type":"string"},"preco":{"type":"number"},"intervaloCobranca":{"type":"string","enum":["MENSAL","TRIMESTRAL","ANUAL"]}}},"PlanoResponseDTO":{"type":"object","properties":{"nome":{"type":"string"},"descricao":{"type":"string"},"preco":{"type":"number"},"intervaloCobranca":{"type":"string","enum":["MENSAL","TRIMESTRAL","ANUAL"]},"id":{"type":"string","format":"uuid"}}},"UserLoginDTO":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"LoginResponseDTO":{"type":"object","properties":{"token":{"type":"string"},"usuario":{"$ref":"#/components/schemas/UserResponseDTO"}}},"UserCadastroDTO":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"cpf":{"type":"string","minLength":1},"password":{"type":"string","maxLength":100,"minLength":6}},"required":["cpf","email","name","password"]},"ProdutoCreateDTO":{"type":"object","properties":{"nome":{"type":"string","minLength":1},"descricao":{"type":"string","minLength":1},"preco":{"type":"number"},"imagem":{"type":"string"},"estoque":{"type":"integer","format":"int32","minimum":1},"descontoEspecial":{"type":"number"},"statusClube":{"type":"boolean"}},"required":["descricao","estoque","nome","preco"]},"ProdutoResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"descricao":{"type":"string"},"preco":{"type":"number"},"estoque":{"type":"integer","format":"int32"},"statusClube":{"type":"boolean"}}},"ItemPedidoDTO":{"type":"object","properties":{"produtoId":{"type":"integer","format":"int64"},"quantidade":{"type":"integer","format":"int32","minimum":1}},"required":["produtoId","quantidade"]},"PedidoDTO":{"type":"object","properties":{"itens":{"type":"array","items":{"$ref":"#/components/schemas/ItemPedidoDTO"},"minItems":1}},"required":["itens"]},"ItemPedidoResponseDTO":{"type":"object","properties":{"produtoId":{"type":"integer","format":"int64"},"nomeProduto":{"type":"string"},"quantidade":{"type":"integer","format":"int32"},"precoUnitario":{"type":"number"},"subtotal":{"type":"number"}}},"PedidoResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"nomeUsuario":{"type":"string"},"itens":{"type":"array","items":{"$ref":"#/components/schemas/ItemPedidoResponseDTO"}},"valorTotal":{"type":"number"},"dataPedido":{"type":"string","format":"date-time"}}},"AssinaturaDTO":{"type":"object","properties":{"planoID":{"type":"string","format":"uuid"}}},"AssinaturaResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"nomeUsuario":{"type":"string"},"nomePlano":{"type":"string"},"precoPlano":{"type":"number"},"dataInicio":{"type":"string","format":"date-time"},"dataFinal":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["ATIVA","CANCELADA","PAUSADA"]}}},"ProdutoUpdateDTO":{"type":"object","properties":{"nome":{"type":"string"},"descricao":{"type":"string"},"preco":{"type":"number"},"imagem":{"type":"string"},"estoque":{"type":"integer","format":"int32"},"descontoEspecial":{"type":"number"},"statusClube":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PagePedidoResponseDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/PedidoResponseDTO"}},"number":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"paged":{"type":"boolean"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Cole o token JWT obtido em /usuarios/login","scheme":"bearer","bearerFormat":"JWT"}}}}