# backend/src/main/java/com/projeto/security/DTOS/user/UserLoginDTO.java
package com.projeto.security.DTOS.user;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record UserLoginDTO(@NotBlank @Email String email, @NotBlank String password) {
}